// src/types/resources/Person.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { THumanName } from '../partials/HumanName';
import { TContactPoint } from '../partials/ContactPoint';
import { TDate } from '../simpleTypes/Date';
import { TAddress } from '../partials/Address';
import { TAttachment } from '../partials/Attachment';
import { TReference } from '../partials/Reference';
import { TPersonLink } from '../partials/PersonLink';

export type TPerson = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  name?: THumanName[];
  telecom?: TContactPoint[];
  gender?: string;
  birthDate?: TDate;
  address?: TAddress[];
  photo?: TAttachment;
  managingOrganization?: TReference;
  active?: boolean;
  link?: TPersonLink[];
};